+++ src/commands/test-raw/index.spec.js
@@ -12,6 +12,7 @@ import testerPluginTmpDir from '@dword-design/tester-plugin-tmp-dir'
 import packageName from 'depcheck-package-name'
 import fs from 'fs-extra'
 import { globby } from 'globby'
+import nodeVersion from 'node-version'
 import outputFiles from 'output-files'
 import P from 'path'
 import unifyMochaOutput from 'unify-mocha-output'
@@ -59,6 +60,21 @@ export default tester(
       },
     },
     empty: {},
+    'global setup': {
+      files: {
+        'global-test-hooks.js':
+          'export const mochaGlobalSetup = () => console.log(1 |> x => x * 2)',
+        src: {
+          'index1.spec.js': 'export default { valid: () => {} }',
+          'index2.spec.js': 'export default { valid: () => {} }',
+        },
+      },
+      async test() {
+        const output =
+          this.base.test() |> await |> property('all') |> unifyMochaOutput
+        expect(output).toMatchSnapshot(this)
+      },
+    },
     grep: {
       files: {
         src: {
@@ -129,7 +145,7 @@ export default tester(
       },
       test() {
         return expect(this.base.test()).rejects.toThrow(
-          'error  Unexpected token o',
+          `error  Unexpected token ${nodeVersion.major === '20' ? "'o'" : 'o'}`,
         )
       },
     },