+++ src/get-generated-files/get-package-config/index.js
@@ -16,6 +16,7 @@ export default function () {
   }
 
   return {
+    type: 'module',
     ...(this.packageConfig
       |> pick([
         'name',
@@ -49,7 +50,7 @@ export default function () {
     ...this.config.packageConfig,
     scripts:
       commandNames
-      |> mapValues((nothing, name) =>
+      |> mapValues((handler, name) =>
         this.packageConfig.name === '@dword-design/base'
           ? `rimraf dist && babel --config-file ${packageName`@dword-design/babel-config`} --copy-files --no-copy-ignored --out-dir dist --ignore "**/*.spec.js" src && node dist/cli.js ${name}`
           : `base ${name}`,